adding imdb dataset for hetero graph

+++ tests/conftest.py
@@ -4,7 +4,7 @@ from arango.database import StandardDatabase
 from arango.http import DefaultHTTPClient
 
 from adbpyg_adapter import ADBPyG_Adapter
-from torch_geometric.datasets import Planetoid
+from torch_geometric.datasets import Planetoid, IMDB
 
 def pytest_addoption(parser) -> None:
     parser.addoption("--url", action="store", default="http://localhost:8529")
@@ -36,11 +36,19 @@ def pytest_configure(config) -> None:
     db = ArangoClient(hosts=con["url"], http_client=NoTimeoutHTTPClient()).db(
         con["dbName"], con["username"], con["password"], verify=True
     )
-
     adbpyg = ADBPyG_Adapter(db)
-
-    dataset = Planetoid("./", "Cora")
-    data = dataset[0]
-    # heterodata = data.to_heterogeneous(node_type_names=['Paper'], edge_types_names=[('Paper', 'Cites', 'Paper')])
-    db.delete_graph("cora", drop_collections=True, ignore_missing=True)
-    adbpyg.pyg_to_arangodb("cora", data, overwrite=True)
+    # loading cora dataset (Homogeneous)
+    dataset_cora = Planetoid("./", "Cora")[0]
+    if db.has_graph('cora'):
+        db.delete_graph("cora", drop_collections=True, ignore_missing=True)
+        adbpyg.pyg_to_arangodb("cora", dataset_cora)
+    else:
+        adbpyg.pyg_to_arangodb("cora", dataset_cora)        
+        
+    # loading IMDB dataset (Heterogeneous)
+    dataset_imdb = IMDB('./imdb')[0]
+    if db.has_graph('imdb'):
+        db.delete_graph('imdb', drop_collections=True, ignore_missing=True) 
+        adbpyg.pyg_to_arangodb("imdb", dataset_imdb)
+    else:
+        adbpyg.pyg_to_arangodb("imdb", dataset_imdb)
